backend/requirements.txt
fastapi
python-dotenv
openai>=1.0.0
pydantic>=2.0
